// src/constant/day-type.ts
export enum DayType {
  regularWorkingDay = 'Regular Working Day',
  restDay = 'Rest Day',
  specialNonWorkingDay = 'Special Non-Working Day',
  specialNonWorkingRestDay = 'Special Non-Working Day and Rest Day',
  regularHoliday = 'Regular Holiday',
  regularHolidayRestDay = 'Regular Holiday and Rest Day',
}